+++ src/wzd_structs.h
@@ -27,6 +27,8 @@
 
 #include "wzd_hardlimits.h"
 
+#include "wzd_types.h"
+
 #include <sys/time.h> /* struct timeval */
 
 /*********************** ERRORS ***************************/
@@ -174,8 +176,8 @@ typedef enum {
 
 /*********************** STATS ****************************/
 typedef struct {
-  unsigned long long	bytes_ul_total;
-  unsigned long long	bytes_dl_total;
+  u_int64_t             bytes_ul_total;
+  u_int64_t             bytes_dl_total;
   unsigned long		files_ul_total;
   unsigned long		files_dl_total;
 } wzd_stats_t;
@@ -198,7 +200,7 @@ typedef struct {
   unsigned short	num_logins;	/**< number of simultaneous logins allowed */
   char			ip_allowed[HARD_IP_PER_USER][MAX_IP_LENGTH];
   wzd_stats_t		stats;
-  unsigned long long	credits;
+  u_int64_t         	credits;
   unsigned int		ratio;
   unsigned short	user_slots;	/**< user slots for gadmins */
   unsigned short	leech_slots;	/**< leech slots for gadmins */
